requirements.txt
agent-framework-github-copilot --pre
python-dotenv>=1,<2